Remove unused eligibility checking protocol
